.claude/skills/run-debate/requirements.txt
google-genai>=1.0.0
openai>=1.50.0
